pyproject.toml
[project]
name="vid2led"
version="0.1.0"
dependencies=[
    "opencv-python",
    "rpi-ws281x-mock"
]